requirements.txt
tensorflow==1.1.0
scipy==1.2.2
nltk==3.4.1
networkx==2.2
tensorboard==1.9.0
distance==0.1.3
click==6.7
multiprocessing
matplotlib==2.2
numpy==1.14.2
scikit-image<0.15
pillow
logging
six
svglib 
